Fir2Ir: fully expand type before converting it
Otherwise, in case of a typealias to an unresolved type in kapt, we
incorrectly skipped the `is ConeErrorType` check and created
`IrSimpleTypeImpl` instead of `IrErrorTypeImpl`. Which later caused the
first check in `IrConstAnnotationTransformer.transformSingleArg` to not
work as intended.

Note that K1 and K2 kapt behavior is different on the test, but K2 kapt
is better because all types must be expanded in Java stubs.

 #KT-79305 Fixed